Remove orphean lot and programme when convention is deleted
