+++ conventions/models/convention_history.py
@@ -1,11 +1,8 @@
 import uuid
 
 from django.db import models
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 
 from conventions.models.choices import ConventionStatut
-from core.services import EmailService, EmailTemplateID
 
 
 class ConventionHistory(models.Model):
@@ -37,47 +34,3 @@ class ConventionHistory(models.Model):
     )
     cree_le = models.DateTimeField(auto_now_add=True)
     mis_a_jour_le = models.DateTimeField(auto_now=True)
-
-
-@receiver(post_save, sender=ConventionHistory)
-def send_survey_email(sender, instance, *args, **kwargs):
-    # send email to get user satisfaction after instructeur validate convention
-    # or bailleur submit convention for the first time ?
-
-    # check if it is the first time the bailleur user submit a convention
-    if (
-        instance.statut_convention == ConventionStatut.INSTRUCTION.label
-        and not ConventionHistory.objects.filter(
-            user=instance.user, statut_convention=ConventionStatut.INSTRUCTION.label
-        ).exclude(id=instance.id)
-        and instance.user.is_bailleur()
-    ):
-        EmailService(
-            to_emails=[instance.user.email],
-            email_template_id=EmailTemplateID.B_SATISFACTION,
-        ).send_transactional_email(
-            email_data={
-                "email": instance.user.email,
-                "firstname": instance.user.first_name,
-                "lastname": instance.user.last_name,
-            }
-        )
-
-    # check if it is the first time the instructeur user validate a convention
-    if (
-        instance.statut_convention == ConventionStatut.A_SIGNER.label
-        and not ConventionHistory.objects.filter(
-            user=instance.user, statut_convention=ConventionStatut.A_SIGNER.label
-        ).exclude(id=instance.id)
-        and instance.user.is_instructeur()
-    ):
-        EmailService(
-            to_emails=[instance.user.email],
-            email_template_id=EmailTemplateID.I_SATISFACTION,
-        ).send_transactional_email(
-            email_data={
-                "email": instance.user.email,
-                "firstname": instance.user.first_name,
-                "lastname": instance.user.last_name,
-            }
-        )
