+++ conventions/services/convention_generator.py
@@ -83,7 +83,6 @@ def _compute_total_locaux_collectifs(convention):
 
 
 def generate_convention_doc(convention: Convention, save_data=False):
-    # pylint: disable=R0912,R0914,R0915
     annexes = (
         Annexe.objects.prefetch_related("logement")
         .filter(logement__lot_id=convention.lot_id)
