+++ conventions/tests/services/test_recapitulatif_service.py
@@ -13,7 +13,7 @@ from conventions.models import Convention
 from conventions.models.choices import ConventionStatut
 from conventions.services import recapitulatif, utils
 from conventions.services.utils import ReturnStatus
-from programmes.models.models import Lot, Programme
+from programmes.models import Lot, Programme
 from siap.exceptions import SIAPException
 from siap.siap_client.client import SIAPClient
 from users.models import User
