+++ programmes/models/__init__.py
@@ -1,2 +1,3 @@
 from .choices import *
+from .indice_evolution_loyer import *
 from .models import *
